+++ src/express.test.ts
@@ -45,4 +45,26 @@ describe('Express integration', () => {
     const body: Result = await response.json()
     equal(body.status, 'OK')
   })
+  test('`path` can be regex', async () => {
+    const app = express()
+    app.use(
+      middleware({
+        path: /hlz.*/,
+        checks: [
+          {
+            id: 'a',
+            fn: async () => 1,
+          },
+        ],
+      }),
+    )
+    const started = app.listen(0)
+    const response = await fetch(
+      `http://localhost:${(started.address() as AddressInfo).port}/hlz123`,
+    )
+    equal(response.status, 200)
+    started.close()
+    const body: Result = await response.json()
+    equal(body.status, 'OK')
+  })
 })