+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Reg. exp. support for custom Express.js path
+
 ### [2.0.0] 2024-05-21
 
 ### Added